Rename fixture files from .client.js
