--- fixtures/flight/src/Counter2.client.js
@@ -1,3 +1,0 @@
-'use client';
-
-export * from './Counter.client.js';
