+++ fixtures/flight/src/Counter2.js
@@ -1,0 +1,3 @@
+'use client';
+
+export * from './Counter.js';
